Add missing meta files imports in worker_app.js
Several meta files of modules that belong to the worker_app were not
imported in worker_app.js when we started migrating extensions out from
their corresponding module.json files. Thus, the functionalities of the
migrated extensions were removed from the worker_app.

This CL adds the corresponding imports to worker_app.js so that the
functionalites are restored.

Fixed: 1180613
Change-Id: I8779a010fa4e37c9593aaaa7e0cf6893b13c6328