[sourcemaps] Support 1:N mappings (inlining / code-splitting).
Previously the DevTools front-end did not support the case where a
single source location corresponds to multiple different locations in
the code that's actually executed by V8 (except in the case of the C/C++
DWARF extension via the DebuggerLanguagePlugin API). This meant that
when a given line in the original source (i.e. in TypeScript) shows up
in multiple locations in the end, the DevTools front-end would only set
a breakpoint in one of these locations and thereby not always pause
correctly.

A good example here is the function inlining performed by JSCompiler,
i.e. consider the following example input:

```js
(function() {
  function absDistance(a, b) {
    return Math.abs(a - b);
  }

  globalThis.manhatten = function(p1, p2) {
    return absDistance(p1.x, p2.x) + absDistance(p1.y, p2.y);
  }
})();
```

JSCompiler will turn that into the following JavaScript by default:

```js
'use strict';
(function() {
  globalThis.manhatten = function(a, b) {
    return Math.abs(a.x - b.x) + Math.abs(a.y - b.y)
  }
})();
```

In particular it inlines the `absDistance` code into the two callsites
within the `manhatten` function to save space (and probably also to
optimize for speed). It will also properly record this fact in the
generated sourcemap. In order to correctly support this optimization,
the DevTools front-end has to set two breakpoints in the `manhatten`
function in the generated code whenever the developer sets a breakpoint
in `absDistance` (via sourcemaps).

The same happens with code-splitting, where a bundler like webpack
"inlines" shared modules in different bundles, and here we also need to
be able to set multiple breakpoints for a single location in the source
code (as reported in https://crbug.com/1142705).

The implementation is surprisingly easy, as the heavy lifting was
already done before when leese@chromium.org added support for multiple
raw locations for the DebuggerLanguagePlugin API. It was mostly a matter
of relaxing the restrictions and also tracking script url/id as part of
the positions (we cannot currently test this, as it needs an orthogonal
fix in the CompilerScriptMappings, and a test case will be added with
that fix).

Also-By: leese@chromium.org
Fixed: chromium:1180794
Bug: chromium:1142705
Change-Id: I7dc0e61f3baf1fd04e021467ff4df7d2ff62b5e3