Fix transparent CodeMirror scrollbar separator
Let's locally revert this change for now and wait for CodeMirror to fix
the issue upstream. I have tried using `inherit` and `initial`, but both
don't work. Since we have a separate darkmode file, we can safely add
the white background back, as in darkmode the other CSS file is loaded.

This was originally applied in CodeMirror in
https://github.com/codemirror/CodeMirror/pull/6485

R=jacktfranklin@chromium.org

Fixed: 1180740
Change-Id: Ia48e3f46ec9e57a2104d30bd9fed523dac1ddf10