[sourcemaps] Add proper support for setting breakpoints in multiple scripts.
With code splitting (a popular feature in bundlers like webpack
and rollup that is supposed to speed up route loading), it can
happen that a shared component / module ends up in scripts for
multiple routes / code splits. In that case DevTools thus far was
unable to set the breakpoint in all relevant locations, but would
only consider the script location from the last loaded sourcemap
that references the source file.

With this change we now properly consider all referencing source
maps for a given source file *and* also lift the restriction in
that a breakpoint would only be set in one raw location (that
is we now set the breakpoint in multiple scripts properly).

Fixed: chromium:1142705
Also-By: pfaffe@chromium.org, leese@chromium.org
Change-Id: Icb43e1e4d2454366545230d4525e895bef417fe3