Make cwd and NODE_PATH absolute in test runner
This CL makes the custom node_modules_path absolute, because Node
requires it to be an absolute path, and does the same thing with the
custom `cwd` argument. Else if you pass a custom `cwd` you'll get errors
in the test suite as running `mocha --config ./foo/config.js` doesn't
work; Mocha also wants the path to be absolute.

Bug: none
Change-Id: Ib626075f4dd1c00fa2998d55fda8c212b8e853c0