Use import type to remove ESLint disable lines
For all of these imports, they were historically only used for
types. Therefore, we can make them `import type * as` to make sure
that the eventual bundle does not include them.

This CL is the result of applying the following regex:

```
import \* as (.+) from '(.+)';\s+// eslint-disable-line no-unused-vars
```

to replace it with

```
import type * as $1 from '$2';
```

for all `*.ts` files.

R=aerotwist@chromium.org

Bug: none
Change-Id: I00ad7bed5d2845a20099d6c13d6eb85ee567f0fd