Animations: Delete line creating extra tabstop
When creating https://crrev.com/c/2491814 ,
I missed a line in the revert. This line was causing an extra tab
stop to be inserted on a div in the animations tool, making it appear
that focus was lost when navigating between nodes. This removes the extra
tabstop.

Repro steps in bug below. Behavior now matches expected.

Bug: 1178428
Change-Id: I30b1360e3e81d95fa7027e8e82624edc4dfe1e07