Retain focus when Navigating To Styles
Whenever the "Navigate to Style" option is selected from the context
menu on an entry in the computed styles pane, a revealer is used
to open and highlight the style in the styles pane that is responsible
for the computed style. However, focus is lost when the revealer
is called.

This change puts focus on the containing section in the styles pane when
the revealer is called.

Before: https://imgur.com/e1L2g1P
After: https://imgur.com/npaVPwb

Bug: 1182417
Change-Id: Ib30574b1e9123f5c5f49fe7ab737fc82ea3951fe