Add third_party lighthouse generated sources to data
The files generated by the third_party lighthouse package were
missing from the data deps. In a release build, this is not an
issue, as the files are rolled up in `front_end/lighthouse` and
`front_end/lighthouse_worker` respectively. However, for legacy
layout tests, we still need these separate files.

R=aerotwist@chromium.org

Fixed: 1183707
Change-Id: I9b04a964a61cf57e01be6f475c23223a08b5561b