Keep Network tab selection between panel switches
If the user prefroms the following steps:
  1) Inspect a Network request
  2) Select a Tab different from 'Header' (e.g 'Timing')
  3) Open Performance (or any other) panel
  4) Switch back to Network panel

Currently the Network panel will show the 'Header' tab instead of the
tab selected in 2). This is due to a wrong assumption in the code
that assumes the View is always freshly constructed. This is not
always the case.

This CL fixes this by force-selecting the tab only once, the first
time the view is shown after construction.

R=sigurds@chromium.org

Bug: chromium:1184578
Change-Id: I444ffacf1bd0d2377af3ab6f9f30482514f1e0c7