Avoid overlaps between paragraphs and links in EmptyWidget
The EmptyWidget is used in a variety of places throughout DevTools to
teach users about what certain areas of the tools do when there isn't
any other information to show.
For instance, when there is no manifest in the manifest section of the
Applications panel, or when there is no previewable response for a http
request.

In the particular case of the manifest, if the space to display the
message is too tiny (e.g. when zooming in), the link may overlap the
text making it hard to see.

This fixes it by getting rid of the column flex layout that was used to
display the bold title, paragraph and link. Those can simply stack up
using the default block flow, and be centered with text align. This
fixes the problem because it forces a scrollbar to appear when space
runs out.

Other instances of the EmptyWidget did not seem to have that particular
problem, but they also don't regress with this change.

Before: https://imgur.com/gN5nfQL
After: https://imgur.com/lS6ns73
Bug: 1185708
Change-Id: I4d43949624fdd8c325c324dfa7006b9a6e61fb86