[pretty-print] Support `import.meta` when pretty printing.
This changes the parsing mode for the JavaScript formatter to 'module',
so that it no longer chokes on `import.meta` references. Generally this
seems to be reasonable and more permissive than defaulting to 'script'.

We could thread throw the module/script information from the engine, but
that would only work for sources that are seen by V8, but not for
scripts or modules that originate from somewhere else.

Fixed: chromium:1185578
Ref: https://github.com/acornjs/acorn/pull/1019
Change-Id: Iabbc51e62b24be28f47124c73c7196ff603787e7