[wasm] Rename "Stack" scope to "Expression".
This is in preparation for the V8 side change, which exposes a dedicated
"stack" object with indexed properties within this scope. The end result
is that (a) "Copy property path" and "Add property path to Watch" will
work as expected for stack slots, and (b) everything that's visible in
the first level of the scope will be valid to enter into the Console for
developers, including "stack", whereas before the "Stack" scope was kind
of an outlier here.

The name "Expression" matches the syntactic concept of the S-expressions
in WebAssembly text format (the stack machine is the corresponding
semantic concept).

Screenshot: https://imgur.com/a/ryFiOGA.png
Bug: chromium:1159310, chromium:1071432, chromium:1159307
Change-Id: I3e02a6f237a03bec98ec826f40d142e4873b2931