Add a slowness factor to the devtools_sanity_unittest
Some tests are timing out on bots that are generally slower than
usual, e.g. msan bots. This CL speculatively adds a slowness factor
to these tests to give them more time. The tests are currently dis-
abled from the chromium side, so once this CL is in, they need to
be activated again, and their flake status needs to be monitored.

Bug: chromium:1185674
Change-Id: Icce2b6e212c2e78a9b7d949773f8d8dff5f536f4