Move accessibility to panels/accessibility
This also updates the Runtime and build_release_applications to use
the correct format for the various files. Most notably, it should
use the basename when concatening file names, since we don't want
to include the `panels/` part in these file names. As such, the
filenames become `panels/accessibility/accessibility_module.js`.

R=aerotwist@chromium.org

Bug: 1187573
Change-Id: I2d3d453ac0e5ea0d4358228531bfe31bde6e7655