[tests] Enable console e2e tests
Wait for the final console evaluation to complete so that we don't
have pending evaluations when we navigate the devtools renderer.
This caused flaky errors after the context group had been destroyed.

I ran a stressor on the trybots with iterations=20 before and after
the fix.

Cq-Include-Trybots: luci.devtools-frontend.try:devtools_frontend_mac_rel
Fixed: 1134579
Change-Id: Ib5fdbc57d8aa18f6a1eb018516cc27f0cac30e00