[TreeOutline] add `expandToAndSelectTreeNode` function
This CL adds a method that takes a tree node and will expand the tree to reveal
that node, and then focus it.

Along the way we ran into async timing issues with coordinator.done() and
LitHtml's `until` directive not being connected; so this CL also includes
improvements to the tests to instead wait until the right number of nodes are
rendered, rather than rely on the fact that they will be rendered in the next
frame, which was working but opens us up to flakey tests.

Bug: chromium:1166669
Change-Id: Ic73f614ae29df3a217524405b81a03ff40c2872b