[sources] Open Watch view when adding expressions.
Previously when using the "Add property path to Watch" from within the
Scope view, there was no visual feedback to the developer unless the
Watch view was already open. With this change, we add explicitly open
and focus the Watch view when adding expressions from the Scope view
(just like we already did for the "source.add-watch" action).

Fixed: chromium:1166607
Change-Id: I2f59387cad86cd13073066e7938797ca9f536eb3