[TestRunner] add flag for configuring DevTools root page.
That. Note that this currently isn't used or set because everyone +
all our scripts still run the Python version of the test runner
script, not the new script. But this needs to be in place before we
move people over to the new script.

We provide a fallback value such that this still works and is able to
run the tests despite the test runner not providing the configuration
via an environment variable.

Bug: chromium:1186163
Change-Id: I8255e5c7f9902be845e3c33add7b1703f6e4db9d