Add distinct shortcut for triggering Sources editors action on mac
These actions didn't have a dedicated shortcut for mac and weren't
triggered in this plaform using the current shortcurs, as they use the
`Ctrl` key (this key is generally mapped to the `Meta` key in mac).
This CL adds a dedicated shortcut for the mac platform to each of
these actions using the `Meta` key.

Bug: none


Fix broken context menu actions in Sources text editor

The context menu of the Sources text editor had been broken as a
regression of https://crbug.com/c/2550688 and it was subsequently fixed
in https://crbug.com/c/2567957 . However, the fix in the latter altered
the original behaviour by relying on CodeMirror's `on` API to listen for
`contextmenu` events instead of the native `addEventListener` API as
originally implemented. This transitively provoked a change in the
flavors added to the application context by the time the actions are
exectued, causing them to not run successfully.

This CL restores the original behaviour of the actions in the context
menu by adding listener via the `addListener` API in the
SourcesTextEditor element.

Fixed: 1189118
Change-Id: Ib060a669c7c9018c5c5462ca0e60017b977262d0