Render Coordinator: await async callbacks
The Render Coordinator currently works its way through all scheduled
read and write callbacks in sequence. However, if any of these callbacks
are async it doesn't await them, which can result in the queue being
marked as empty too early.

This CL updates the Render Coordinator to await all read and write
handlers (which is a noop for non-Promise handlers), and only when each
is completed will it move to the next handler.

R=jacktfranklin@chromium.org

Bug: 1189614
Change-Id: I90e110bebaa4e012bd02ea3b108dfd1ec9180826