Write the ability to view the computed value of a CSS property in the context menu
Added a new entry in the context menu when
righ-clicking a CSS property in the Styles tab.
Choosing it will show the computed value
of just that CSS property in the Computed tab.
This behaviour is a fast and easy way to filter a
specific CSS property from all that are shown
in the Computed tab.

Bug: 1076198
Change-Id: Ib985bfa49659e64fd1941a82a71763d77a4a1b03