[LMI] Fix responsiveness bug
When resizing the LMI, at some point the byte view would not correctly
resize, leading to the LMI using a scrollbar to be able to show the
entire LMI.

The bug was caused by a miscalculation of 1 pixel, as one of the used
variables for calculating how many bytes to show in one row is
rounded, while the others are not.

This now accounts for the rounding done for clientWidth.

Fixed: chromium:1189575
Change-Id: I1d1bd8f406e78b0f80605539937ae3a0ceb4fad1