Fix incorrect background color for selected item in quick open menu
The dark theme color for this element was forced to be the same as its
text's in hrrps://crrev.com/c/2773812, as a consequence there is no
contrast between the element's background and its text.

This CL deletes the whole `:host-context(.-theme-with-dark-background)`
as we don't need overrides anymore per the dark theme migration. Also,
for better contrast, the background color was changed to a different
variable.

Before: https://imgur.com/a/qegZaWD
After: https://imgur.com/a/m4WLB5i

R=jacktfranklin@chromium.org

Bug: chromium:1152736
Change-Id: I4679ef7fdcefa102fdb473fb6a7fb5ba09d1eb64