Change inspectorCommon css variable for better contrast
The value previously set for the `--accent-color` variable in
inspectorCommon.css provides low contrast between background and text in
highlighted elements in DevTools, when running in dark mode. This CL
changes the value of this variable for another one declared in
themeColors.css for better contrast.

Before: https://imgur.com/a/y5arsVS
After: https://imgur.com/a/SbhIBug

Bug: chromium:1152736
Change-Id: Ibb6f7d0c6e5b82c99816edcc9ac785dcd7074c63