[l10n] Don't pass progress stats for all arrays in _parseArray
We were trying to pass a placeholder with stats in it to all UI
strings from _parseArray, but we call it in two different places
with UI strings that don't have placeholders.

If the input array does not have a length provided then we
can't provide progress updates and just need to update the
status.

This bug was causing the i18n library to throw at runtime
with the 'placeholder not found' error.

Bug: none
Change-Id: I8a377e82c344e8175c11d1a06c8f68840450b012