[Breadcrumbs]: fix scroll buttons not being removed
This CL fixes a bug where if the breadcrumbs were overflowing and then
the user updated them such that they could all fit in the container,
the scroll icons were not removed. This was due to a bug in our logic
that checked for overflows, it only detected the case where we go from
overflow: false and need to set overflow: true. It did not detect the
opposite, where we were overflowing but now are not.

Note that this CL also reminded me that the breadcrumbs are not
migrated to use the new Component Render coordinator. I will do that
in a follow up CL (logged bug: https://crbug.com/1190788).

Fixed: chromium:1189139
Change-Id: I067d45ebc99f39f63ab40f288cc919b2f6f1f349