Add loadTestModule for legacy test runners
This prepares the legacy test runners to be moved into a sub-folder.
Doing so will ensure that the test runners only load the relevant
JavaScript files. If the implementation-under-test relies on CSS files,
then the layout tests should import the relevant implementation modules
themselves.

After the move, we can update the implementation in `loadTestModule` and
we don't have to update the layout tests themselves again.

R=aerotwist@chromium.org

Bug: 1187573
Change-Id: I183fe4e5ca6be29ac4b77a9f04fca5585e384734