Fix empty Styles pane issue when no style is matched for css variable
When there is no matched style for a css variable declaration, the
Styles pane rendering would throw an error, which causes an empty Styles
pane. This CL creates a plain text node instead of throwing in this
case, allowing the rest of the Styles pane to render as expected.

Fixed: chromium:1190406
Change-Id: I4d63b9430a64462dc76a981b06f4ddad5e828233