Fall back to `Page.getResourceContent()` when loading resources.
Previously we'd only try `Network.getResponseBody()`, when trying to
load the content of a request, but the inspector only has a limited
cache size and the resource payload might have been evicted from that
cache at some point. `Network.getResponseBody()` will then generate an
error, saying that they body is no longer there.

In that case we will now fall back to `Page.getResourceContent()` to
(re)request the resource payload from the server, and display that to
the developer. We already use that method when we don't have the request
in the first place (i.e. when DevTools was opened while the page was
already loaded).

Fixed: chromium:1185198
Also-By: pfaffe@chromium.org
Change-Id: I6ed2f52aa4e2378f1edbe32757004d6088560f6e