Enable font editor tests and fix flakiness
The tests were relying on the order in which styles are rendered
in the styles panel. It appears that the order is not guaranteed at
least at some points in time. This CL updates the font editor
tests to click on a font editor button that exists in the element.style
section instead of clicking on any button that is found first.

Bug: chromium:1184658
Change-Id: I75de95462dd2d9dab1a4d99112685f682bbe80e0