Remove NetworkRequest._requestCookies
NetworkRequest._requestCookies parses cookies included with a request
manually in the frontend. This is not strictly necessary, as all
included cookies are available via
NetworkRequest.(included/blocked)RequestCookies().

The only remaining use-site of NetworkRequest._requestCookies was the
HARLog. There, we simply combine the included AND blocked cookies
to properly log the actual made request.

R=sigurds@chromium.org

Bug: None
Change-Id: I28f2d54dc4e4b7a0771e76116d0b0a093da94345