Fix deps in release mode for devtools_entrypoint
In release mode, the devtools_entrypoint prebundle step was missing
the deps. We ran into this issue when we were moving the emulation
panel into panels/emulation, which required some updates to the
toolbox.ts entrypoint (crrev.com/c/2782540).

The missing deps then required the rootdir logic to be changed,
such that deps are correctly resolved relative to the gen-directory,
not the source directory.

This CL might improve build performance in release builds, since
we are now properly using incremental references for our
entrypoints and saving a bit of compilation time.

DISABLE_THIRD_PARTY_CHECK=TypeScript fix
R=aerotwist@chromium.org

Bug: 1187573
Change-Id: I4662ac90ba50116cb668dfe3164fe725b74f9a37