[quick-open] Improve outline for HTML files.
This adds explicit outline support for HTML files, which were previously
parsed via acor-loose, which led to _questionable results_. This new
version now explicitly searches for <script> and </script> tags and
sends the text in between to the JavaScript outline. It's still very
basic (i.e. doesn't work well if <script> occurs inside of "strings" or
<!-- comments -->, but that seems quite unlikely anyways, and would
require significantly more than 15 lines of easy to read code to deal
with).

Drive-by-fix: Unify CSS outline reporting with the JavaScript and HTML
outline handling. We should be able to improve this further by also
looking for <style> tags in the HTML and invoking the CSS outline
generator in that case.

Fixed: chromium:1191465
Bug: chromium:761019
Change-Id: I9e66a438e120f7a7df98a6ad049d1266d82c68bf