Permissions Policy: show details for disallowed features
This CL adds a 'Show Details' button to the Permissions Policy section
in a frame's details view. When clicked, for each disallowed feature
DevTools shows a link to the/a reason why this feature is disallowed via
Permissions Policy. This can be either to the corresponding request's
headers in the Network panel or to the frame's element in the Elements
panel.

Video: https://i.imgur.com/2S9PFvM.mp4

Bug: 1158827
Change-Id: I4b4db8d7e34e5886fff48264ee8c528a351725b7