Add Media Feature Support querying to `common`
This patch lifts the media feature support check function from
`inspector_main` into `common`, allowing it to be used to remove
Command Menu options when a media feature is not supported.

Currently `prefers-reduced-data` takes advantage of support checking
because it's only supported on Mac - there is currently a bug on Windows
where the option to emulate appears in the Command Menu. In the future
`forced-colors` emulation will utilize this function, since users
can disable support in `about://flags`

Bug: 1096068
Change-Id: I8e01abed0f9bf9c037f068c17690317089eec376