Re-enable codemirror.darkmode.css
This CL re-enables the use of the generated dark mode file for
cmdevtools.css, after the bugs in the script generation were fixed
(see bug for details).

I have verified that the bugs caused by the dark mode CSS file
previously are now resolved and don't occur with the new CSS file
generated with the script fixed.

Bug: chromium:1188524
Change-Id: Ie997dcf6ee1ea8c03b2e5c7fb5d03febfee3c55b