Create a base mocha config file
Whilst doing the test runner work over the last couple of weeks I
noticed that we have a variety of `mocharc.js` files that are 99% the
same, and I've had to make the same changes to most of them. This is
the first step at creating a base function that can generate the parts
of the config that are the same for each test suite.

Bug: chromium:1186163
Change-Id: I818b6831b8953084ee667f44c35a8f279fbdc55f